Moved machine learning code to separate file (not imported by default)
